requirements.txt
Flask==2.2.5
google-generativeai==0.0.9
